first draft: beamer article and new templates
For scripted talks, we can use the beamerarticle functionality. Working
on that makes clear that pandoc templates would be a better solution
than having a zillion TeX files that need to be copied around all the
time. So moving a bunch of stuff into those templates.

We'll setup handouts while we're at it.

And control everything from a single Makefile.

scripts/script.md is the new sample for a written out talk.

Some sporadic biber/biblatex glitches still.