overlay_filter update for pandoc >= 1.18
pandoc 1.18 changed the JSON format and therefore broke the tree-walking
routine I had taken from the old pandocfilters.

https://github.com/jgm/pandoc/releases/tag/1.18

pandocfilters has been updated and I have again lifted the code from the
updated version as of

https://github.com/jgm/pandocfilters/commit/debaac6fa32ea377441940e9db8c9b28ce34674d

simplifying the routines there (which allow for a sequence of "actions"
on the tree instead of just the one).

Since I don't want to require users of this repository and scuro to have
pandocfilters installed, this is the best solution I can come up with.

On the basis of some superficial spot-checking, I believe this is still
compatible with pandoc <= 1.17 as well.