generate bibliography slides if biblio-title given
The code is the same as in default.beamer, except that beamer's styling
is a little over-the-top. The preamble removes what beamer.pdf calls
"cute little icons" before each bib item and adjusts the continuation
text for the slide header.

A bibliography slide with no header requires tweaking the template.