options for ::: noslide ::: delineator, if any
some feature bloat for noslide.lua. If the output format is beamer,
noslide Divs are removed. Otherwise, they are kept but surrounded on
both sides by an element. By default this is a Null, but it can also be
something else, e.g. a rule. The choice is controlled by metadata
variable `noslide`.