Back on track with rule-delimited slides
Lots of changes and wafflings to get here.

noslide.lua eliminates the elaborate "delineator" options and hard-codes the horizontal rules. But I add another filtering pass to collapse two HRs in a row into one. We still make the lecture script with output format beamer (and we set a metadata variable hide_noslide in the Makefile to modify what the lua filter does). The beamerarticle template now puts a rule at the start of each frame. Because the HRs are treated as slide delimiters in beamer output, what we end up with is rules where we want them, delimiting each slide and each section of script (which is technically a "slide" for article purposes).

Updated script-sample.md to reflect the new workings.